refactor: add stake trade param
