+++ packages/trader/src/Modules/Trading/Components/Playground/new-trade-params-container.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 // import classNames from 'classnames';
 import { ButtonToggle } from '@deriv/components';
 
-const NewTradeParamsContainer = ({ onClick }: { onClick: () => void }) => {
+const NewTradeParamsContainer = ({ onClick }: { onClick: (trade_param: string) => void }) => {
     return (
         <div className='trade-param_section'>
             <div className='trade-param_section_title'>
@@ -35,15 +35,15 @@ const NewTradeParamsContainer = ({ onClick }: { onClick: () => void }) => {
                 <div className='trade-param_name'>Multiplier</div>
                 <div style={{ fontSize: '16px', lineHeight: '24px' }}>x15</div>
             </div>
-            {/* <div className='trade-param_container'>
-                <div className='trade-param_name'>Barrier</div>
-                <div style={{ fontSize: '16px', lineHeight: '24px' }}>+1.00</div>
-            </div> */}
-            <div className='trade-param_container'>
+            <div className='trade-param_container' onClick={() => onClick('stake')} onKeyDown={() => onClick('stake')}>
                 <div className='trade-param_name'>Stake</div>
                 <div style={{ fontSize: '16px', lineHeight: '24px' }}>10.00 USD</div>
             </div>
-            <div className='trade-param_container' onClick={onClick} onKeyDown={onClick}>
+            <div
+                className='trade-param_container'
+                onClick={() => onClick('risk_management')}
+                onKeyDown={() => onClick('risk_management')}
+            >
                 <div className='trade-param_name'>Risk management</div>
                 <div style={{ fontSize: '16px', lineHeight: '24px' }}>Take profit: 1.00 USD</div>
             </div>
